cypress update concurrent end-to-end tests

+++ test-end-to-end/webapp-cypress/run-cypress-test.sh
@@ -6,10 +6,8 @@ exit_with_error() {
   exit 1
 }
 
-#Parse command line arguments as per
-# /usr/share/doc/util-linux/examples/getopt-example.bash
-TEMP=$(getopt -o 'r' --long 'spec:' -n 'run-cypress-test.sh' -- "$@")
-
+# Parse command line arguments using getopt
+TEMP=$(getopt -o 'r:' --long 'spec:,testservercontrollerUrl:,learningspaceurl:' -n 'run-cypress-test.sh' -- "$@")
 eval set -- "$TEMP"
 unset TEMP
 
@@ -17,72 +15,78 @@ unset TEMP
 WORKDIR=$(pwd)
 SCRIPTDIR=$(realpath $(dirname $0))
 SPEC=""
-CONTROLSERVER=""
+TESTSERVER_URL="http://localhost:8075/"   # Default value
+LEARNING_SPACE_URL="http://localhost:8087/"  # Default value
 
-
-
-
+# Parse the arguments
 while true; do
-        case "$1" in
-          '-r'|'--spec')
-              echo "Set spec to $2"
-              SPEC=$2
-              shift 2
-              continue
-              ;;
-         '--')
-                        shift
-                        break
-                ;;
-
-	esac
+    case "$1" in
+        '-r'|'--spec')
+            echo "Set spec to $2"
+            SPEC=$2
+            shift 2
+            continue
+            ;;
+        '--testservercontrollerUrl')
+            echo "Set testserver Controller URL to $2"
+            TESTSERVER_URL=$2
+            shift 2
+            continue
+            ;;
+        '--learningspaceurl')
+            echo "Set Learning Space URL to $2"
+            LEARNING_SPACE_URL=$2
+            shift 2
+            continue
+            ;;
+        '--')
+            shift
+            break
+            ;;
+        *)
+            break
+            ;;
+    esac
 done
 
-if [ "$CONTROLSERVER" = "" ]; then
-  CONTROLSERVER="http://localhost:8075/"
-fi
-
 # Ensure the results directory exists
-if [ ! -e $SCRIPTDIR/results ]; then
-  mkdir $SCRIPTDIR/results
+if [ ! -e "$SCRIPTDIR/results" ]; then
+    mkdir -p "$SCRIPTDIR/results"
 fi
 
-if [ -d "$SCRIPTDIR/results" ]; then
-  # Check if there are files to delete
-  if [ "$(ls -A $SCRIPTDIR/results)" ]; then
+# Clear the results directory if it contains files
+if [ -d "$SCRIPTDIR/results" ] && [ "$(ls -A $SCRIPTDIR/results)" ]; then
     rm -rf "$SCRIPTDIR/results"/*
 fi
 
-fi
-
-
-# Start control server
-$SCRIPTDIR/../../testserver-controller/start.sh --siteUrl http://localhost:8087/ || exit_with_error "Failed to start control server"
+# Start the control server
+echo "Starting control server with site URL: $LEARNING_SPACE_URL"
+$SCRIPTDIR/../../testserver-controller/start.sh --siteUrl "$LEARNING_SPACE_URL" || exit_with_error "Failed to start control server"
 
 # Define the spec argument
-SPECARG=$SPEC
-if [ "$SPEC" != "" ]; then
-  SPECARG="$SCRIPTDIR/cypress/e2e/$SPEC.cy.js"
+if [ -n "$SPEC" ]; then
+    SPECARG="$SCRIPTDIR/cypress/e2e/$SPEC.cy.js"
 else
-  SPECARG="$SCRIPTDIR/cypress/e2e/"
+    SPECARG="$SCRIPTDIR/cypress/e2e/"
 fi
 
-cd $SCRIPTDIR
+# Navigate to the script directory
+cd "$SCRIPTDIR" || exit_with_error "Failed to change directory to $SCRIPTDIR"
 
 # Run npm install and Cypress tests
+echo "Running 'npm install'..."
 npm install || exit_with_error "Failed to run 'npm install'"
-#Note additional -- required to split up arguments between npm exec and cypress
-# as per https://stackoverflow.com/questions/63854769/cypress-runs-all-test-files-even-when-spec-parameter-is-used
-npm exec cypress run -- --spec "$SPECARG"  || exit_with_error "Cypress test run failed"
+
+# Run Cypress tests
+echo "Running Cypress tests..."
+npm exec cypress run -- --spec "$SPECARG" || exit_with_error "Cypress test run failed"
 
 # Capture the exit status of the Cypress test run
 teststatus=$?
 
 # Stop the control server
+echo "Stopping control server..."
 "$SCRIPTDIR/../../testserver-controller/stop.sh" || exit_with_error "Failed to stop control server"
 
 # Exit with the same status as the Cypress test run
 exit $teststatus
-
-
-
